requirements.txt
python-fasthtml
mysql-connector-python
fastapi
numpy
scipy
folium
matplotlib